fixed up the html & css (still using custom elements)
what was there was, like, the fasted possible version--no head or metadata.

added all that. and made sure it works on mobile (horizontal & vertical).

not sure if there is a better way to do the media queries. like, i want
to put something in that groes inversely to the vh shrinking. i'm sure
there is a way to do that with calc()... but i don't know it/don't wanna
fuck around with that right now.

i just put in more breaks for now. good enough.

-z.